requirements.txt
# Core Framework
fastapi==0.118.0
# Avoid Windows build issues from extras (watchfiles/httptools); install plain uvicorn
uvicorn==0.34.2
python-multipart==0.0.18

# Database
sqlalchemy>=2.0.40,<3
alembic==1.14.0
aiosqlite==0.19.0
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.4.0
# Note: ecdsa@0.19.1 (from python-jose) has known vulnerabilities but no fix available yet
# SNYK-PYTHON-ECDSA-6184115, SNYK-PYTHON-ECDSA-6219992
passlib[bcrypt]==1.7.4
bcrypt>=4.2.1,<5
# cryptography 43+ publishes cp313 wheels; avoids building Rust
cryptography>=44.0.0,<45
python-dotenv==1.0.1

# Data/time utils
python-dateutil==2.9.0

# Background task scheduling
APScheduler==3.10.4

# AI/ML
openai>=1.0.0
openai-chatkit

# Document Processing
pypdf==6.1.1
# Document processing (core)
python-docx>=1.1.2,<2
# Google Cloud Vision API for document OCR and extraction
google-cloud-vision>=3.8.1,<4
pdf2image>=1.17.0
pillow>=10.4.0

# Reports/Exports
weasyprint>=63.1
XlsxWriter>=3.2.0
openpyxl==3.1.2

# (Optional analytics libs moved to requirements.optional.txt)

# API Documentation
pydantic>=2.10.6,<3
email-validator==2.2.0
pydantic-settings==2.7.1

# Pin Starlette to latest stable (security fixes)
starlette==0.47.2

# Development & Testing
pytest==8.3.4
pytest-asyncio==0.25.2
httpx==0.28.1
requests==2.32.3
black==24.10.0
flake8==7.1.1

# Slack integration
slack_sdk>=3.34.0,<4

# Google Calendar integration (optional; guarded by ENABLE_GCAL)
google-api-python-client>=2.158.0,<3
google-auth>=2.38.0,<3
google-auth-httplib2>=0.2.0,<1
google-auth-oauthlib>=1.2.0,<2
pytz>=2024.1

# Monitoring
prometheus-client==0.21.1
structlog==24.4.0

# Additional utilities
click==8.1.8
rich==14.0.0

# Animation Engine for Learning Module (3Blue1Brown-style lessons)
manim>=0.18.1,<1
# Required for Manim rendering
pycairo>=1.27.0
# FFmpeg required for video rendering (install via system package manager)

# SEC Data Ingestion for Learning Module
sec-edgar-downloader>=5.0.3

# PDF Signing (Free alternative to DocuSign)
reportlab>=4.2.5
PyPDF2==3.0.1
# Optional heavy/analytics libs (install only if needed)

# Data processing
pandas>=2.2.3,<3
numpy>=2.1.1,<3
openpyxl>=3.1.2

# Analytics/DB
duckdb>=1.1.3,<2

# Image processing (for OCR doc extraction)
pillow>=10.4.0

# Document processing (optional)
pdfplumber>=0.11.0
pymupdf>=1.24.10
paddleocr>=2.7.0
invoice2data>=0.4.0